Refine dataprep microservice code structure
