--- comps/dataprep/langchain/qdrant/docker/docker-compose-dataprep-qdrant.yaml
@@ -1,39 +1,0 @@
-# Copyright (c) 2024 Intel Corporation
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#    http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-version: "3"
-services:
-  qdrant-vector-db:
-    image: qdrant/qdrant
-    container_name: qdrant-vector-db
-    ports:
-      - "6333:6333"
-      - "6334:6334"
-  dataprep-qdrant:
-    image: opea/gen-ai-comps:dataprep-qdrant-xeon-server
-    container_name: dataprep-qdrant-server
-    ports:
-      - "6000:6000"
-    ipc: host
-    environment:
-      http_proxy: ${http_proxy}
-      https_proxy: ${https_proxy}
-      QDRANT: ${QDRANT}
-      QDRANT_PORT: ${QDRANT_PORT}
-      COLLECTION_NAME: ${COLLECTION_NAME}
-    restart: unless-stopped
-
-networks:
-  default:
-    driver: bridge
