--- comps/dataprep/langchain/qdrant/requirements.txt
@@ -1,15 +1,0 @@
-beautifulsoup4
-docarray[full]
-easyocr
-fastapi
-huggingface_hub
-langchain
-numpy
-opentelemetry-api
-opentelemetry-exporter-otlp
-opentelemetry-sdk
-Pillow
-pymupdf
-qdrant-client
-sentence_transformers
-shortuuid
