+++ comps/dataprep/redis/requirements.txt
@@ -1,0 +1,16 @@
+beautifulsoup4
+docarray[full]
+easyocr
+fastapi
+huggingface_hub
+langchain
+langchain-community
+numpy
+opentelemetry-api
+opentelemetry-exporter-otlp
+opentelemetry-sdk
+Pillow
+pymupdf
+redis
+sentence_transformers
+shortuuid
